resolved comments

+++ src/services/src/Demoulas.ProfitSharing.Common/Contracts/Response/BeneficiaryInquiry/BeneficiarySearchFilterResponse.cs
@@ -4,7 +4,7 @@ namespace Demoulas.ProfitSharing.Common.Contracts.Response.BeneficiaryInquiry;
 public record BeneficiarySearchFilterResponse
 {
     public int BadgeNumber { get; set; }
-    public int PsnSuffix { get; set; }
+    public short PsnSuffix { get; set; }
     public string? Name { get; set; }
     public string? Ssn { get; set; }
     public string? Street { get; set; }
